move compose RPC method to independent advanced namespace
