+++ src/rpc/BatchRequester.js
@@ -1,17 +1,20 @@
 const RPCError = require('../provider/RPCError');
 
 const SEND_TX_METHOD = 'cfx_sendTransaction';
+const SEND_RAW_TX_METHOD = 'cfx_sendRawTransaction';
 
 class BatchRequester {
   /**
    * BatchRequester constructor.
    *
-   * @param {Provider} provider - A provider instance
+   * @param {Conflux} conflux - A Conflux instance
    */
   constructor(conflux) {
     this.conflux = conflux;
     this.requests = [];
     this.decoders = [];
+    this.accountNextNonces = {};
+    this.accountUsedNonces = {};
   }
 
   /**
@@ -46,12 +49,40 @@ class BatchRequester {
     });
   }
 
+  _markNonceUsed(from, nonce) {
+    if (!this.accountUsedNonces[from]) {
+      this.accountUsedNonces[from] = {};
+    }
+    this.accountUsedNonces[from][nonce] = true;
+  }
+
+  isNonceUsed(from, nonce) {
+    return this.accountUsedNonces[from] && this.accountUsedNonces[from][nonce];
+  }
+
+  async _getNextNonce(from) {
+    let _nonce = this.accountNextNonces[from];
+    if (!_nonce) {
+      _nonce = await this.conflux.advanced.getNextUsableNonce(from);
+    }
+    //
+    while (this.accountUsedNonces[from] && this.accountUsedNonces[from][_nonce]) {
+      _nonce += BigInt(1);
+    }
+
+    this.accountNextNonces[from] = _nonce + BigInt(1); // update next nonce
+
+    return _nonce;
+  }
+
   /**
-   * Clean Batch requester's requests and decoders
+   * Clear Batch requester's requests and decoders
    */
-  clean() {
+  clear() {
     this.requests = [];
     this.decoders = [];
+    this.accountNextNonces = {};
+    this.accountUsedNonces = {};
   }
 
   /**
@@ -61,7 +92,6 @@ class BatchRequester {
    */
   async execute() {
     // prepare transaction nonce and sign it
-    const accountNonces = {};
     const _requests = [];
     // eslint-disable-next-line no-plusplus
     for (let i = 0; i < this.requests.length; i++) {
@@ -69,18 +99,18 @@ class BatchRequester {
       if (req.method === SEND_TX_METHOD && req.params[0] && req.params[0].from && this.conflux.wallet.has(req.params[0].from)) {
         // prepare nonce
         const from = req.params[0].from;
-        let nonce = accountNonces[from];
-        if (!nonce) {
-          nonce = await this.conflux.cfx.getNextUsableNonce(from);
-          accountNonces[from] = nonce;
+        let _nonce = req.params[0].nonce;
+        if (!_nonce || this.isNonceUsed(from, _nonce)) {
+          _nonce = await this._getNextNonce(from);
+          req.params[0].nonce = _nonce;
         }
-        req.params[0].nonce = nonce;
-        accountNonces[from] += BigInt(1);
+        this._markNonceUsed(from, _nonce);
+
         // sign transaction
         const account = await this.conflux.wallet.get(`${from}`);
         const signedTx = await account.signTransaction(req.params[0]);
         // change method to cfx_sendRawTransaction
-        req.method = 'cfx_sendRawTransaction';
+        req.method = SEND_RAW_TX_METHOD;
         req.params[0] = signedTx.serialize();
       }
       _requests[i] = req;
